+++ djmoney/forms/widgets.py
@@ -9,7 +9,9 @@ __all__ = ('MoneyWidget',)
 
 class MoneyWidget(MultiWidget):
 
-    def __init__(self, choices=CURRENCY_CHOICES, amount_widget=None, currency_widget=None, *args, **kwargs):
+    def __init__(self, choices=CURRENCY_CHOICES, amount_widget=None, currency_widget=None, default_currency=None,
+                 *args, **kwargs):
+        self.default_currency = default_currency
         if not amount_widget:
             amount_widget = TextInput
         if not currency_widget:
@@ -20,7 +22,7 @@ class MoneyWidget(MultiWidget):
     def decompress(self, value):
         if value is not None:
             return [value.amount, value.currency]
-        return [None, None]
+        return [None, self.default_currency]
 
     # Needed for Django 1.5.x, where Field doesn't have the '_has_changed' method.
     # But it mustn't run on Django 1.6, where it doesn't work and isn't needed.